Some modelling of DRM events
Reading from a file descriptor representing a DRM card returns a series of
event objects. These types and utilities are to help with decoding those
event objects and producing more usable safe representations of them.

This also includes some basic support for "pageflip" calls, because that's
one way to cause an event to be delivered in order to test all of this.