Enough property-wrangling to do atomic modeset
It's expensive to individually fetch the metadata for each property name
from the card, so the new functions here are intended to allow a one-time
property enumeration on startup and then just using the u32 property ids
directly after that.

Overall the idea here is to allow the atomic commit to be the only ioctl
request made once per frame, by getting everything else done ahead of time
in the setup code.

The modesetting-atomic example here isn't doing quite the right thing in
how it's deciding which plane to use to drive each crtc. We'll hopefully
improve on that in future commits, but for now this is enough to get a
successful atomic modeset on a simple system that has only one active
connector that already has the kernel's console running on it.