Card::open accepts any impl AsRef<CStr>
In practice it'll be common for callers to allow their users to specify
which card device to use, which will then typically appear as a CString
rather than just a &CStr. This signature is therefore more convenient
because it allows both, and also allows Cow<CStr> as is now used in our
examples to allow specifying a card device on the command line.